regions: a native tail callee takes a cycle member by-move
The closure-cycle merge refused a cycle whose letrec body tail-calls a
non-member with a member as a direct argument. The refusal's argument is a
callee that replaces the frame: its new activation owns the member as a
parameter and releases it, decrefing the merged arena a second time against the
deferred release. A native callee does neither. It borrows its arguments and it
keeps the frame, so the binding-scope DecrefRegion stays live and is the arena's
single release.

Record on the binding whether its compile-time constant is a native function —
from the value, so a primitive-scope name core.lisp rebinds to a bytecode
closure reads as a closure — and ask may_replace_frame before refusing.

The everyday factory reaches this: a letrec of mutually recursive closures over
shared state, handed back as a struct literal in tail position, which compiles
to a tail call on the struct native over both members. It leaked its whole
arena per construction.

Both files the change touches were past the reading budget, so the letrec
body's tail reading and the analyzer's ambient-environment binders each move to
a file of their own, and the merge's restatement of letrec.md gives way to the
gate list a reader of the function needs.
